OMP::Translator: create 2SB image subsystems
This adds a new method "create_image_subsystems" which copies
the existing subsystem information.  Its operation is controlled
by a new configuration parameter "auto_image_subsys_2sb".
The "bandwidth_mode" method call has been moved just before
this new method is called.

The "correlator" has been adjusted again so that it now iterates
directly over subbands, then receptor and finally enters a loop for
the required number of CMs.  This avoids the need to construct a
list of SPW IDs with repeated entries and makes it easier to
handle the allocation of image subsystems to already-allocated
receptor slots and LO2s.