OMP::Translator: calc. mask after frontend config
Since the last step of both the ACSIS and SCUBA-2 frontend config
methods was to set up the receptor or subarray mask, this can
instead be performed in the main JCMT "translate" method immediately
after the frontend has been configured.  This will allow the
"calc_receptor_or_subarray_mask" method (and in turn the
"tracking_receptor_or_subarray" method which it uses) to access
the frontend configuration.