OMP::Translator: remove rest freq. adjustment
Now simply set the requested first subsystem IF frequency as the
instrument's IF center frequency, rather than adjusting the rest
frequency to include the offset.

The previous calculation did not appear to take redshift into
account, so it presumably did not work correctly when there was
an IF offset and non-insignificant redshift.

This update assumes that it will be possible for the instrument
software to be updated to read the IF center frequency from the
configure XML (overriding any value read from the initialize
XML).  If this cannot be done, this commit should be reverted
along with the previous commit making the translator use the
OT's IF value (7e8fab5164).  It might be possible instead to use
the OT's value for the sideband flip, but the instrument initialize
XML's value for the offset.  If rest frequency offsetting is
reinstated, it would be a good idea to take redshift into consideration.