OMP::Translator: alter JOS calculations for FRSW
For long integration times (greater than the "freqsw_max_seq_length"
configuration parameter), the previous calculation had the effect of
rounding the observation up to a multiple of that parameter so as to
match or exceed the requested time.  Thus times up to 10 minutes were
translated correctly, but all times between 10 and 20 minutes would be
rounded up to 20 minutes.

The calculation has been changed so that we first calculate the
number of cycles required, taking into account the configured maximum.
Then the number of steps is calculated to approximate the requested
observing time, given the number of cycles.