OMP::MSBDB: avoid "magic" remaining counters
The OMP::MSBDB::_insert_row method uses OMP::Info::MSB::summary
to get a hash containing data to be inserted into the MSB table.
However this method sets "magic" remaining counter values "REM"
and "N/A" for removed and undefined values.

With our current database, this causes an error "Incorrect integer
value" when storing a science program containing new MSBs with
a removed remaining counter.

I'm not sure if the magic values are still required given apparent
changes to how removed MSB counters work, or whether the hash form of
the summary output could always be returned before they are applied.
For now the safest way to address the inability to store such science
programs seems to be to explicitly disable the magic values in this
particular case.