admin/ompchecktarget: Change --semester option
The "--semester" option has been replaced with a new option "--scheduled"
which works similarly in that it selects scheduled observations (from the OMP)
rather than completed observations (from the JCMT database).  However it is
no longer necessary to specifiy a semester -- all JCMT projects will be
searched excluding those in the EC, JAC and EAO queues.