MSBs: add notification on projects first accepted msb of night
This is for JCMT only. The MSBDB::MSBDone function has an option
notify_first_accept that defaults to 0, and even if it is 1 will
not send any emails unless the telescope is JCMT.

MSBServer::MSBDone will always set this to 1.

If the msbtitle is passed into the MSBServer::MSBDone function then
this will populate through. Else the msbtitle will be shown as
'UNKNOWN'.