client/sql_client: Check for user / host arguments
These previously worked, as arguments given to mysql after those
generated by the script overrode them, but the "Connecting to ..."
message still showed the default values.  Now the script checks
for a -u / --user / -h / --host command line option and uses it
to replace the user or host name obtained from the configuration.