OMP::Translator::JCMT: remove scan offset rotation
We previously rotated the offset of a scan through an angle of
$info{OFFSET_PA} - $info{MAP_PA}.  It is unclear why this was done!
Perhaps it was to aid in setting up a mosaic of offset scans.  However
the offset and scan could have different systems, so this would not
always work!  Furthermore the OT always set the two angles to the
same, so the difference would be zero and no rotation would be applied!

This linkage is being removed from the OT, so it should be removed
here also before it starts triggering unexpectedly.