OMP::Auth: add new user authentication module
This module is intended to replace the use of project-based passwords
by authenticating users individually.  It allows for the use of
different authentication providers, with the first being staff access
via LDAP.  The modules is primarily designed for used from the web
interface to the OMP, via the CGI module.  However it also allows
for direct use of a username and password where necessary.