OMP::CGIPage::Fault: redirect after updates
Now after filing / updating / responding to a fault, redirect back
to the fault viewing page.  This helps avoid duplicate submissions
without needing "form keys".

In this case of printing, a message is shown instead.  This is because
we don't currently have a way to show a message after the user returns
to the viewing page, and it should be made clear that the fault has
been printed.