OMP::Translator::ACSIS: config. rotator slew opt.
The rotator slew option has been made configurable.  This replaces
the previous logic in the 'rotator' method which seems like it would
never currently trigger.  The mode can be specified in general or
by mode, but initially it has been configured to use the current
LONGEST_TRACK option.  The intention is to allow for experimentation
via local.cfg.