OMP::Translator::ACSIS: alter LO2 comments & log
This updates the comments in the code related to LO2 settings to
better match what seems to be happening.  Previous references to
"channel 1" were replaced with "first channel" (because the calculations
look like they may instead operate for channel 0).  The log output
for "RefChanIF" was renamed to "FirstChanIF" because the IF for the
reference channel should be fixed, whereas this value isn't.  The
sketches showing hybrid subsystem layouts have been reordered to
account for numbering the channels from the high-IF end downwards.