OMP::Translator: try to "flip" about default IF
For instruments with asymmetric IF bandwidths, mirroring about the
default IF is not guaranteed to produce subsystem IF bands which are
within the IF bandwidth.  In order to preserve the IF of subsystems
at the default IF, we now attempt to perform the sideband "flip"
about the default IF.  However, before doing so, we need to check
whether the subsystems will fit, and if not, apply a suitable offset
to the mirroring frequency.  Since we do not have the details of
the IF bandwidth available, new configuration entries have been added
to give the range for instruments for which it is asymmetric.