web: Adjust schedule staff columns
This changes the staff columns to have more generic names so as not
to imply anything about whoever might be listed in that column.
The end date has been added.  The columns were rearranged to keep
night and morning operator columns next to eachother and place the
end date next to the morning operator (since that is the morning
they will be operating).