web: Allow access to schedule entries as iCalendar
This adds a new database table "ompschedcal" which configures
calendars to be exported from the telescope schedule.  These have
a name and a pattern to be matched against the staff columns
on the schedule.  They also each have a secret token to allow
calendar applications to retrieve the calendar entries without
needing to log in.

Currently there is a page to list all the calendars (not yet
linked from anywhere) but the calendars have to be configured
manually.