web: Replace index.html with a dynamic home page
This home page makes use of a new template parameter of the write_page
method, so that its content method returns a template context hash
rather than printing the page directly.  The index template uses a new
layout wrapper templated intended to be used instead of the current
HTML::WWWTheme module.  To use this new home page, the web server
will need to be configured to use cgi-bin/index.pl as the index,
e.g. by the DirectoryIndex directive.