web: Configure red theme for "dev" version
Previously the omp-dev installation was modified locally to set a
red theme.  This commit adds some of this theming, based on
settings in the omp-dev.cfg file.  A new favicon based on the
OMP logo is included, with a red version for the "dev" environment.
Multiple stylesheets can now be specified in the configuration files.