OMP::MSBDB: extract _obs_row_to_coord function
This extracts from the MSB querying code (_run_query method)
the creation of an Astro::Coords object from the information
in the observation table.  This allows this operation also
to be performed for the getSciProgInfo method, so that coordinates
are included in its result, allowing it to be used in more
cases where previously fetching the whole science program
was required.