OMP::ObslogDB: allow updateObsComment on time gaps
This did not tend to work for time gaps since they do not have
an OBSID.  Therefore a function to create the placeholder values
was extracted from _reorganize_comments and used to try to match
time gaps to retrieved comments.  Additionally, time gap objects
may not have startobs defined, so when sorting, fall back to
endobs in this case.