web: Use template for data retrieval page
This required collecting messages which were previously
printed to stdout during the PackageData process, including from
constructing an OMP::Info::ObsGroup.  Some code which was duplicated
between the _package_data and _package_data_cadc subroutines
has been moved to the main request_data subroutine.

Note that currently only retrieval via CADC is supported,
since this is used for all JCMT data, and this script is running
on the OMP web server rather than the main Hilo web server.
The package creation method may work, but will not be ideal
with the template approach as nothing will be shown until the entire
process is complete, including the message asking the user
to wait while their packages are created.  If this method
is to be enabled again in future, it would be better to move
it offline, so that it doesn't hold up the web server.  I.e.
it could write a request to the database, return a web page
saying that packaging is in process, and then a server
on the Hilo network with the data available could create the packages.