web: Use templates for observation log pages
To aid the transition to temnplates, the main body of these pages,
listing the observations, is currently provided to the templates
as HTML captured with Capture::Tiny.

Loading the night report with the "delta" option (for a multiple-night
report) previously generated a report ending the day before the current day,
running back "delta" (-1) days from then.  This has been simplified to
ending on the current day, so that switching between the single- and
multiple-night reports keeps the same end date.

Some unavailable weather plots have been commented out for now.