OMP::Info::ObsGroup: use placeholder obsid sub.
When locating time gaps, now use the OMP::ObslogDB::_placeholder_obsid
subroutine to generate placeholder obsid values to match with the
associated comments.  This replaces duplicate logic in this routine
with a call to the recently-extracted subroutine.

It looks like the first time this was being done in locate_timegaps
was not necessary as the database query would have already used
the _placeholder_obsid subroutine to fill missing obsid values.