OMP::FaultGroup: call timelost when needed
The timelostTechnical and timelostNonTechnical methods both,
if called without arguments, returned information which
would have been cached in the object by a prior call
to the timelost method.  However this relied on the order
in which these methods were called.

This commit alters those methods so that, if the corresponding
entry is undefined, timelost is called to prepare the required
information.

(This is not still entirely foolproof, e.g. if timelost had
previously been explicitly given a time, the new no-argument
call would not then cause it to recompute all times, including
the information required for these functions.  However it should
help in the case of OMP::NightRep where timelostbyshift gets a
new OMP::FaultGroup from its faultsbyshift method, so timelost
will not have been called.)