OMP::CGIComponent::Obslog: extract get_obs_summary
This extracts a method "get_obs_summary" from the "obs_table"
method to prepare all of the information to be shown in the
table.  This allows the text output mode to be split out into
a separate method "obs_table_text".  Both methods use the
information prepared by the new summary method.