web: Use template for observation lists
This completes the conversion of the night report to templates.
The OMP::NightReport::ashtml and OMP::CGIComponent::Obslog::obs_table
methods have been removed.  The new version does not (yet) support
WORF links, but this is not a problem as WORF does not (yet) support
current JCMT instruments, and we do not have any files in /ukirtdata.

The observation status labels were previously copied to OMP::Info::Obs.
This seems like a better place for this information, so the display
classes and display order have been moved there too.

Templates are now rendered with access to "private" attributes,
since OMP::Info::Obs::nightlog returns entries such as "_ORDER"
to which the templates need access.