OMP::Translator::JCMT: override pt/foc for raw XML
Override the pointing and smu offsets in Raw XML observations
with the current values from the wiredir's instrument XML.
This is to allow Raw XML to be used without the need to update
the MSB whenever the pointing model or focus offsets change.
However it would be better for this to be optional in case
it is ever desired to use Raw XML observations to test a
pointing model.  (Therefore it is applied in an "if (1)"
block, where a suitable condition can be applied later.)