OMP::Mail::Original: Apply encoding to email data
The body of messages should now be encoded as UTF-8.  The headers
should be encoded with a version of the MIME-Header encoding
modified to avoid wrapping.  (Otherwise the headers can be wrapped
twice as Mail::Header also does this.)  The line folding length
has been increased to 318 in order to avoid folding since our
mail server apparently has problems with this.