OMP::CGIPage: add decoded_url_param method
This decodes (as UTF-8) the parameter value from the CGI module's
url_param method.  (CGI appears only to decode parameters fetched
via the param and multi_param methods.)

So far this is applied to the projectid, faultid and category
parameters which are added as arguments to view methods from the
corresponding CGIPage modules.