OMP::FaultUtil: add format_fault opt. max_entries
The format_fault method now accepts optional hash-style options,
with the previous third argument (base URL) becoming the first
such option, "omp_url".  A new option "max_entries" can be used
to limit the number of responses included in the formatted output.
If truncation occurs (i.e. max_entries is less than the number
of "responses") then the header information (fault details and
original filer information) is included either at the start or
end as appropriate.  (Otherwise it continues to be included with
the first "response", i.e. the original filing, as before.)

Additionally the code of this method was tidied slightly, with the
aim of making it easier to see where '<br>' tags should and
should not be included.