Translator::ACSIS: try to read red_recipe_id value
Now use the "_tree" attribute of the RedConfigList object
read from the ACSIS recipe file to try to find the value for
red_recipe_id.  If it is not found, use "unknown version".
This replaces the previous behavior of always setting this
attribute to "incorrect. Should be read from file".