templ: Show single shift time acct. if present
Now if time accounting information is available for a single
shift, show that shift instead of the overall information.  This
is to match the behavior of the OMP::NightRep::astext method.
It should ensure that comments are always visible, since these
are not currently included in the overall information.

(cherry picked from commit 5b51facbeedeeba4f35b4c142b62bff3de416c7a)