web: Update user details page
The page design has been refreshed, using a table for the basic
details and a combined table of projects with capacity filter buttons.
The details editing page now has a heading to be more consistent.

The OMP::CGIPage::render_template method now adds a "json" vmethod
so that the template can easily generate the data attribute required
for filtering.

See also fault: 20230802.004