Info::Obs: add subsystems_tiny for non-retainhdr
The 'subsystems' method can only be used when the "retainhdr"
option was enabled on construction so that full FITS headers
are available.  This commit adds a new private attribute
'_subsys_tiny' which is filled by the '_populate' method when
not retaining headers.  This uses similar grouping logic
as the 'subsystems' method to store information about the
differences found in the subsystems.  The new 'subsystems_tiny'
method then applies these differences to copies of the info
object to provide a similar interface to the main 'subsystems'
method.

Note: the new functionality (via tiny subsystem information)
has been kept separate from the main 'subsystems' method to avoid
disruption to that method, which is used elsewhere (e.g. in jsasubmit
from the JSA module).  It is intended to be a lighter-weight
alternative for display purposes (e.g. in the obs. log) without
adding too much to the size of the associated arccache files.
If this approach turns out to be successful, then perhaps the
main 'subsystems' method could instead have an option to allow
fall-back to tiny subsystem information.