templ: Show multiple subsystems in obs. log
To try to make the log more readable when there are multiple
subsystems, the first few columns span multiple rows.
For the "Run" and "UT" columns this includes spanning
subsequent comment rows.  (Comments are now shown in separate
rows of the table.)  Time gaps are now shown with "Gap" in
the "Run" column and the UT (start) time in the "UT" column
(assuming these are the first two columns).  To save space
the comment link is shown below the status label.

See also: fault 20080804.002