Info::MSB: change wavebands to Astro::Waveband
This is to allow Astro::WaveBand to format values for display.

The generated accessor was renamed to '_wavebands' (with type
@Astro::WaveBand) to allow a custom method 'wavebands' which
can gather the waveband objects from constituent observations
if needed.

The 'waveband' accessor still returns a concatenated string
(to retain the previous behavior) but if given a value,
it must be an Astro::WaveBand object.

OMP::MSBDoneDB attempts to guess the "units" of waveband values
in the 'ompmsbdone' table in order to construct Astro::WaveBand
objects.  A new test script was added for this.