CGIComponent::Helper: move url_absolute to CGIPage
The url_absolute function has been moved to a method of the
OMP::CGIPage class.  As this was the last remaining function
in OMP::CGIComponent::Helper, this module has been removed.
Methods of OMP::CGIPage and OMP::CGIComponent objects can
now access this method as $self->url_absolute or
$self->page->url_absolute.  For the OMP::Auth modules,
the OMP::CGIPage object is now given instead of the
plain CGI object, so that they can access this method.