CGIPage::Fault: add "Automatic" response filtering
This adds to the "All" and "Non-hidden" response options.  This could
become the default in future, but retain "Non-hidden" as the default
for now to allow the new mode to be tested.

This mode is intended to show a given number of non-hidden messages at
the start and end (currently 12 each) and then any messages
flagged above "normal" in the middle of the thread.