Display: add new format_html method
This method takes either preformatted HTML or plain text and
prepares it for display as HTML.

It has been applied to the OMP::FaultUtil->format_fault method.

Note: this method makes use of the existing "preify_text" which
itself attempts to detect preformatted text via a leading <html>
tag.  This check must be removed once the preformatted flag is
enabled everywhere.