Mail: add preformatted option to build method
This is to indicate whether the given message is HTML or not.  If it
is preformatted HTML, a separate plain text version of the message
body can also be provided.  A new test script has been added to
check the three modes of operation (plain text only, HTML only and
HTML with a plain text version).