FaultDB: suppress excess wrapping of fault emails
Since in this case, OMP::FaultUtil->format_fault will have already
wrapped the text, suppress wrapping by OMP::Mail.  Note that this
is an unusual case for the OMP because format_fault includes styling
tags inside <pre> sections so we must not wrap the HTML code.
Typically the OMP handles preformatted HTML with <pre> sections
including text which still needs to be wrapped.