DB::MSB: distinguish unremove from undo method
This adds a method "unremoveMSB" which is separate from the "undoMSB"
method (matching the pair of methods found in OMP::MSBServer).

The OMP::MSB->undoObserve method no longer unremoves removed MSBs,
just returning false in this case.  Now OMP::MSB->unRemove must be
called instead.

A new msgtype constant OMP__FB_MSG_MSB_UNREMOVED is used to log
unremoval in the feedback table.

The logic for fetching an MSB from a science program, updating it in
some way and storing the science program back to the database has
been extracted into a wrapper OMP::DB::MSB->_apply_msb_operation.
This allowed the addition of the new "unremoveMSB" method without
duplicating this logic.  This wrapper is also used by the "alldoneMSB"
method which performs a similar task.