Info::ObsGroup: time accounting trace option
This adds an option "trace_observations" to the projectStatsSimple
method.  If specified then the method will also collect details of
the observations contributing to each project's calculated
time accounting.  This information is shown in a new section
at the end of the time accounting page.

The previous format (positional) argument was changed to an
option "by_shift" to allow the addition of this new option.