SciProgStats: also work with OMP::Info::SciProg
This changes OMP::SciProgStats so that, instead of extending
OMP::SciProg, it simply provides a method which can be used with
either an OMP::SciProg or OMP::Info::SciProg object.

(The interface of OMP::MSB and OMP::Info::MSB doesn't quite match
so for now check which we have to determine whether to call an
"obssum" or "observations" method.  It might be better to instead
make these interfaces more similar.)