QStatus: extract group_queue_status function
This function contains the grouping functionality which previously could
optionally be performed by the query_queue_status function.  Extracting it
into a separate function allows the main query_queue_status function to
always return the same type of information.  The new "extra" attribute of
OMP::Info::MSB is used to attach the times at which the MSB is available
to the returned objects, and group_queue_status accumulates these values
to produce the project hits per hour lists.