templ: Make fault response date link to obs. log
The fault response date is now a link to the corresponding UT date's
observation log page, if the fault category is a telescope and
editing is allowed (to exclude the project fault pages).  The same
criteria have been applied to the "actual time of failure" so that
it only shows as a link when appropriate.