templ: Combine obs. log time summary into 1 table
This uses a new help method of OMP::NightRep,
"get_time_summary_combined", to get the time accounting
summary for all shifts combined.

The fixed width of the table was removed to allow it to expand
as needed and the "hrs" units were removed from the times to
save horizontal space.  Otherwise the table should function
as before.