CGIPage::Project: redirect from home page if CR
On the project home page, trap the UnknownProject error getting
the project details.  When this happens, do a reverse continuation
request search, to look for the corresponding original project.
If found, redirect to that project's home page.  Otherwise show
a not-found error page.

Note this will only work when logged in with staff access as
project authorization will not work for continuation requests.