DB::Shift: add separate updateShiftLog method
The "enterShiftLog" method now always attempts to add a new
entry to the database, while "updateShiftLog" uses the "id"
attribute of the given OMP::Info::Comment object to determine
which entry should be updated.  This is intended to simplify
updating shift log entries, avoiding the need to provide
matching date, user and telescope values.