web: Improve shift entry submission via page
This allows the time zone to be selected (UT or HST).  The time
and text are filled in the form if present (to allow the page to
potentially be updated to re-display the entered data in case
of error).  Removed unnecessary "zone" parameter.  Allow seconds
to be omitted.  If time is not specified, do not fill it in
until the comment is being prepared for storage in the database.