web: Add new "replace_omp_links" filter
This is used in place of the previous sequence of ".replace()"
calls when displaying fault text.  It calls a new method
OMP::Display->replace_omp_links to perform the substitutions,
passing any hash-style options to that method.  If a "fault"
option is given then substitute links to fault images and
response numbers.  Otherwise just substitute links to
fault pages for fault IDs.  The filter is now also applied
to shift log entries (where it will currently only replace
links to fault IDs).