CGIPage::PkgData: add 'direct' retrieval method
This is to allow individual files to be downloaded directly
from the OMP server.  In this mode a list of observations is
displayed, with each having a sub-list of links to raw data files.
When a file is requested, it will be sent with a content type of
application/octet-stream.

This is intended for use by specific observing projects where the
project members need to access data in raw time, e.g. for comet
projects.  It is restricted to only recent data (currently one
week).  It should in addition be restricted to specific projects.