Translator::ACSIS: restore Namakanui slot mapping
This restores the temporary slot mapping for Namakanui of commit
6443affcf4, but now controlled by the lo2_synth_status settings in the
configuration files.  It currently uses a basic test: if the first
subsystem is not possible with LO2 #1 or #2 then it swaps the slot
ordering so that the LO2s will be allocated in the order #3, #4, #1, #2.

The existing HARP slot mapping now also checks the lo2_synth_status
settings, only performing the swap if the subsystems would use different
synthesizers and one of the required high synthesizers is not available,