Translator::ACSIS: search possible slot mappings
Instead of applying a few slot mappings for common cases, make a list
of all possible slot mappings for the requested tuning.  Then try
these in order until one is found which meets the restrictions
imposed by the "lo2_synth_status_low" and "lo2_synth_status_high"
configuration parameters.  If none of the possible mappings
can be used, report an error "Unable to find a slot mapping where
the required LO2 synthesizers are available".  (This message does
not specify which synthesizer / frequency caused the failure
since multiple combinations may have been tried.)

A new "Correlator configuration" section was added to the
translator output to give details of the slot mapping process,
any mappings which were tried but found not to be possible given
the current LO2 status and finally a list of the configuration
of each correlator module.