FaultUtil: use OMP::Display->replace_omp_links
This replaces a specific function "_make_fault_id_link" which
this module used to use to replace fault identifiers with links.
This will ensure the same formatting in emailed fault messages
as on the web interface, including the new links to specific
responses.

The "replace_omp_links" function gained an option "complete_url"
for use in the case of email messages.  (It is called "complete"
rather than "absolute" as the OMP already uses that term to
refer to path-absolute URLs, with full path but no protocol
or host.)  It now uses the OMP configuration system to get
the OMP URL (if needed) and the path to the CGI directory.