Info::ObsGroup: include bad obs orig. project IDs
In projectStatsSimple, if the "trace_observations" option is given,
include in the comments a note of the original project ID of
bad observations.  To do this, record the original project ID
before changing to the bad observation project ID.  Also include
this in the list of gaps.  We only add project information
to this list when the project ID has changed, so the original
project ID will only be recorded correctly for the first
bad observation after a gap.  However this should suffice
since we currently only look at the very next entry in the
list when processing gaps.