client/affiliationstats: Add --bydate option
This script has been updated to use Getopt::Long for option processing.
A new option "--bydate" was added to change the behavior of the script.
If specified, instead of querying for all projects matching the given
semester, it looks up the boundaries of the given semester and searches
for time accounting records in this range of dates.

Additionally the projects without affiliation are sorted by time
observed and name of their PI is shown.  Avoid division by zero
when computing the final affiliation percentages.