DateSun: make get_freeut_range method public
Removed the leading underscore from the name of this method and
changed its documentation from a comment to POD.  Changed the
date parameter name from "start" to "date".

Changed OMP::QStatus to use this method.  (It previously got the
"freetimeut" from the configuration system and then passed it to
another private method, _process_freeut_range, but that is what
this method does, so it might as well use it to perform both steps.)