DB: add "result cache" to database objects
This adds a new attribute "ResultCache" and accessor method
"_result_cache".  They are named "result cache" to avoid confusion
with the more specific "cache" present in OMP::DB::Archive objects.
As an added convenience the "_result_cache" method takes an optional
key argument and will then return the associated value from the cache,
initializing it to a hash reference if it does not already exist.
An additional method to clear the cache could be added if needed.