Fault: rename to "combined" faultLocation method
This renames the "faultLocation_Safety" method to "faultLocation"
to match the other methods where the category is now given as a
parameter (or taken from the object instance) rather than being part
of the method name.  The category is retained in the interface even
though it is not currently needed.  Since this method takes options,
undef must be passed first if not specifying a category, and so "uc"
must only be applied if the category was not given as undef.  Added
POD for the method.