CGIPage::Fault: pass fault obj. to handler methods
In the "_write_page_extra" method, if a fault ID is given, we
retrieve the whole fault (including responses) in order to determine
the category.  Therefore in order to avoid the handler methods having
to fetch the same fault again, add the fault object rather than the
fault ID to the extra handler arguments.