Fault::Group: correct clearing of categories
Passing undef to the "categories" method is supposed to clear
the categories list -- it must restore it to being an empty
array rather than undef since at the end of the method it
is used as an array reference.  Also do this to reset the
list when storing a new collection of faults, as is done
for time lost.