DB::Fault: do not fetch proj. in getAssociations
The getAssociations method is used to search for faults associated
with a particular project.  Therefore we can assume that the user
is interested in that project and not retrieve the list of projects
associated with each fault found.  This should speed up this method
since retrieving the project lists currently involves a separate
database query for every fault found.