NightRep: simplify timelostbyshift method
Instead of taking an argument to indicate which type of time loss
to compute (technical, non-technical or total), return hashes
containing all three types.  This should be more efficient as
the method needs to construct the set of fault groups by shift
type, so it might as well extract all three measures at once.