SOAPServer: move compressReturnedItem method here
OMP::MSBServer and OMP::SpServer both had a version of this method.
This moves a combined version to the OMP::SOAPServer base class.
The combined version returns a SOAP::Data object if
$ENV{'HTTP_SOAPACTION'} is present but does not return
a value indicating whether the output was zipped or not.
The separate "_find_return_type" method from OMP::MSBServer
was retained to allow this to be tested separately.