Query::TimeAcct: include project table in query
This is to allow queries which include a constraint on the telescope
name.  We also retrieve the telescope name since it is now available.
Note that the join to the project table is performed as a "left outer
join" since not all time accounting records will match an actual
project in the database.