Project::TimeAcct: move summary method to Group
This moves the OMP::Project::TimeAcct->summarizeTimeAcct class method
to an instance method "summary" of OMP::Project::TimeAcct::Group
which automatically works on the records contained within the group.

The method now returns a hash reference (instead of the hash itself)
and should ensure that all entries are already Time::Seconds objects.