Query::Archive: adj. dates in _post_process_hash
Move the adjustment of date ranges from the "_create_sql_recurse"
method to the "_post_process_hash" method.  It looks like the
"_create_sql_recurse" method from the base class was copied
here in commit d0e0393924 with only this aspect altered.
Moving the adjustment to the "_post_process_hash" method
should eliminate the only significant difference between
the "_create_sql_recurse" methods in this class and the
base class.

Added a couple of "sort" operations to the query elements so
that the SQL representation of a query can be checked in a test
script.