Query: simplify handling of like expressions
Now that none of the query classes use "like" in place of a full
text search (previously OMP::Query::Archive included the functionality
but it was never used), remove the processing of "like" expressions
to emulate text search.  This allows "like" queries to simply
specify "like" as the comparator for the "_querify" method rather
than having to use a different name of "explicitlike".

Added a "sort" operation to the query elements so that the SQL
representation of a query can be checked in a test script.