NightRep: store MSB done records in the object
This renames the existing "msbs" method to "msbs_by_project"
and adds a more basic "msbs" method which can be used to either
store or retrieve an array of OMP::Info::MSB objects.  When this
method retrieves records from the database, it does so with the
"transactions" option so that we get all comments for each
MSB transaction ID.  This allows for the addition of a "historyMSBtid"
method which can be used in place of the corresponding database
method but which will obtain the information from the C<OMP::NightRep>
object's cache of MSB records.