Project::TimeAcct::Group: store projects in object
This adds an accessor method "project_details" which can be
used to fetch the details of all (non-special) projects in
the time accouting records.  The details are retrieved from
the database when first requested and then stored in the
object.  This method is then applied in the "_get_accts"
method to simplify the querying for project information
and avoid repetition of such queries (e.g. if the "ec_time"
and "science_time" methods are called).