Avoid using JAC::Setup from libraries
This is to allow scripts and applications to configure library paths
as desired.  Include the setup in the "cgi" and "server" initialization
scripts to ensure consistent library loading here.  (Hopefully this
will avoid the case where Astro::FITS::HdrTrans has been loaded but
is unable to load any of its translation classes.)