Info::Obs: check for empty SUBHEADERS
When preparing the "_subsys_tiny" information, skip over empty
hashes in the "SUBHEADERS" list.  (These can be present where there
is only one set of headers.)  Also avoid an undefined value warning
if the subsystem key is not present (or is undefined) by providing
a default value.

Added a test script to check that the "subsystems_tiny" method operates
correctly in the cases with and without "SUBHEADERS".