Query::Archive: remove custom _qhash_tosql method
This was identical to the base class method part from working on
individual sections of the query hash.  We can simply use the
base class method if we arrange for it to see one section at
a time.  (Unfortunately since it doesn't accept a hash parameter
we have to do this by locally setting the QHash attribute.)