client/racover: Add --plot command line option
This can be used to specify a filename, either with .pdf or .png
extension.  If not specified, the plot is displayed in an X window.

Changed the default device from xwin to xcairo.  The "save_plot"
function did not seem to work.  Therefore only generate the plot
once, to the specified file (if any) rather than attempting to
display to X as well as saving to a file.