Info::Obs: remove pol/fts_in from '_subsys_tiny'
When preparing generic headers for '_subsys_tiny', remove headers
which have likely been inferred incorrectly.  E.g. INBEAM will probably
be the same in all subsystems so the generic headers POLARIMETER and
FOURIER_TRANSFORM_SPECTROMETER will be false since INBEAM is missing.
In order to allow 'subsystems_tiny' to reconstruct the 'pol_in' and
'fts_in' attributes correctly, these must be allowed to be undefined.
Convert to 'T' and 'F' strings only when preparing for display.