Translator::ACSIS: ability to ch. Uu p/f to 5GHz
This adds to the check for Uu standards at 6GHz IF an additional check
for Uu pointing or focus at 6GHz.  If either high synthesizer 0 or 1
and either 2 and 3 (0-indexed) is unusable (such that a slot re-ordering
will not suffice for a pointing/focus in chained mode) then change the
IF to 5GHz in the same way as for standards.